test_requirements.txt
bok-choy==0.7.1
ddt==1.3.1
mock==3.0.5
pycodestyle==2.6.0
pylint==2.4.2
pytest==4.6.6
pytest-cov==2.8.1
pytest-django==3.7.0
selenium==3.14.0
xblock-sdk==0.2.2
pypng==0.0.20
